Make upload-x86-jtag target flash only x86 core
This script originated from PVT-User flashpack which was intended to run after 
PVT-Factory flashpack. Since PVT-Factory flashpack put test ARC binary into ARC 
core, it was important for PVT-User flashpack to put a blank ARC binary to reset
the board into a clean state.

This script is now called as a part of make upload-x86-jtag target. In this 
context, it makes no sense to also flash ARC core.